requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
supabase>=2.6.0
fastapi>=0.111.0
uvicorn>=0.30.0
pyTelegramBotAPI>=4.14.0

